+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `nosection` flag to `task create`
 - Add `No section` option to section selection
+- Add `no_sections` config option
 
 ## 2023-09-04 v0.5.1